# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: ATM Booth Banking API
  version: 0.0.1
  description: REST backend for the ATM Booth full-stack project (InHolland). Demo users are seeded on startup — see README.
servers:
  - url: http://localhost:8080
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List customer users
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CustomerStatus'
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - name: sort
          in: query
          required: false
          description: Sort expression such as `createdAt,desc`
          schema:
            type: string
      responses:
        '200':
          description: Page of customer users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerSummaryPage'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Employee role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get customer user details by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer user detail with accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetailResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Employee role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [Users]
      summary: Update customer profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequest'
      responses:
        '200':
          description: Updated customer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'
        '400':
          description: Validation error or invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Employee role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AccountType:
      type: string
      enum: [CHECKING, SAVINGS]
    AccountStatus:
      type: string
      enum: [ACTIVE, CLOSED]
    CustomerStatus:
      type: string
      enum: [PENDING, ACTIVE, CLOSED]
...
    UserRole:
      type: string
      enum: [CUSTOMER, EMPLOYEE]

    # ── Requests ─────────────────────────────────────────────────────────────
...
    CustomerUpdateRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/CustomerStatus'
        firstName:
          type: string
          maxLength: 64
        lastName:
          type: string
          maxLength: 64
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{10,15}$'
        absoluteTransferLimit:
          type: number
          minimum: 0
          description: Applied to both accounts on first activation. Defaults to 1000.00 if omitted.
        dailyTransferLimit:
          type: number
          minimum: 0
          description: Applied to both accounts on first activation. Defaults to 500.00 if omitted.
...
    AccountSummaryResponse:
      type: object
      properties:
        iban:
          type: string
        type:
          $ref: '#/components/schemas/AccountType'
        balance:
          type: number
        status:
          $ref: '#/components/schemas/AccountStatus'
...
    CustomerSummaryResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        createdAt:
          type: string
          format: date-time
    CustomerDetailResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        bsn:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/CustomerStatus'
        createdAt:
          type: string
          format: date-time
        totalBalance:
          type: number
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummaryResponse'
    CustomerProfileResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        bsn:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/CustomerStatus'
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
...
    CustomerSummaryPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CustomerSummaryResponse'
...
    PageMetadata:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        pageable:
          $ref: '#/components/schemas/PageableMetadata'
        sort:
          $ref: '#/components/schemas/SortMetadata'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
    PageableMetadata:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        unpaged:
          type: boolean
        sort:
          $ref: '#/components/schemas/SortMetadata'
    SortMetadata:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean